backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
sqlalchemy[asyncio]==2.0.36
aiosqlite==0.20.0
alembic==1.14.0
python-jose[cryptography]==3.3.0
bcrypt==4.2.1
pydantic[email]==2.10.3
pydantic-settings==2.6.1
python-multipart==0.0.18
openai>=1.54.0
